+++ CHANGELOG.md
@@ -479,3 +479,5 @@ All notable changes to Project Genesis will be documented in this file.
 - Implemented `is_sand_dancer` trait. Entities with this trait gain energy when they are inside a sandstorm event or standing on sand during a storm.
 
 - Implemented `is_water_dancer` trait. Entities with this trait gain energy when on `water` terrain.
+
+- Implemented `is_forest_dancer` trait. Entities with this trait gain energy when on `forest` terrain.